trying to handle precision

+++ includes/libftprintf.h
@@ -6,7 +6,7 @@
 /*   By: vyudushk <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/06/20 16:07:48 by vyudushk          #+#    #+#             */
-/*   Updated: 2017/06/29 15:24:44 by vyudushk         ###   ########.fr       */
+/*   Updated: 2017/06/30 01:10:30 by vyudushk         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -30,7 +30,7 @@ typedef struct	s_flag
 	int		percent;
 	char	type;
 	int		dot;
-	int		pres;
+	size_t	pres;
 }				t_flag;
 
 typedef struct	s_length
